aa module rename for consistency
